domain_test_service::domain

Struct DomainNode

pub struct DomainNode<Runtime, RuntimeApi>
where
    Runtime: DomainRuntime,
    RuntimeApi: ConstructRuntimeApi<Block, FullClient<Block, RuntimeApi>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: ApiExt<Block> + Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block> + MessengerApi<Block, NumberFor<Block>, BlockHashFor<Block>> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, Runtime::AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + RelayerApi<Block, NumberFor<Block>, NumberFor<Block>, BlockHashFor<Block>>,
{

    pub domain_id: DomainId,
    pub key: Runtime::Keyring,
    pub task_manager: TaskManager,
    pub client: Arc<FullClient<Block, RuntimeApi>>,
    pub backend: Arc<Backend>,
    pub code_executor: Arc<RuntimeExecutor>,
    pub network_service: Arc<dyn NetworkService>,
    pub sync_service: Arc<SyncingService<Block>>,
    pub addr: MultiaddrWithPeerId,
    pub rpc_handlers: RpcHandlers,
    pub operator: DomainOperator<RuntimeApi>,
    pub tx_pool_sink: TracingUnboundedSender<ChainMsg>,
    pub base_path: BasePath,

}

A generic domain node instance used for testing.

Fields
domain_id: DomainId

The domain id

key: Runtime::Keyring

The node’s account key

task_manager: TaskManager

TaskManager’s instance.

client: Arc<FullClient<Block, RuntimeApi>>

Client’s instance.

backend: Arc<Backend>

Client backend.

code_executor: Arc<RuntimeExecutor>

Code executor.

network_service: Arc<dyn NetworkService>

Network service.

sync_service: Arc<SyncingService<Block>>

Sync service.

addr: MultiaddrWithPeerId

The MultiaddrWithPeerId to this node. This is useful if you want to pass it as “boot node” to other nodes.

rpc_handlers: RpcHandlers

RPCHandlers to make RPC queries.

operator: DomainOperator<RuntimeApi>

Domain oeprator.

tx_pool_sink: TracingUnboundedSender<ChainMsg>

Sink to the node’s tx pool

base_path: BasePath

The node base path

Implementations

impl<Runtime, RuntimeApi> DomainNode<Runtime, RuntimeApi>
where Runtime: Config<Hash = H256> + Config + DomainRuntime + Send + Sync, Runtime::RuntimeCall: Dispatchable<Info = DispatchInfo, PostInfo = PostDispatchInfo> + Send + Sync, <<Runtime as Config>::OnChargeTransaction as OnChargeTransaction<Runtime>>::Balance: Send + Sync + From<u64> + FixedPointOperand, u64: From<BlockNumberFor<Runtime>>, RuntimeApi: ConstructRuntimeApi<Block, FullClient<Block, RuntimeApi>> + Send + Sync + 'static, RuntimeApi::RuntimeApi: ApiExt<Block> + Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, <Runtime as DomainRuntime>::AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + MessengerApi<Block, NumberFor<Block>, BlockHashFor<Block>> + RelayerApi<Block, NumberFor<Block>, NumberFor<Block>, BlockHashFor<Block>> + OnchainStateApi<Block, <Runtime as DomainRuntime>::AccountId, Balance>, <RuntimeCallFor<Runtime> as Dispatchable>::RuntimeOrigin: AsSystemOriginSigner<<Runtime as Config>::AccountId> + Clone,

pub fn wait_for_blocks(&self, count: usize) -> impl Future<Output = ()>

Wait for count blocks to be imported in the node and then exit. This function will not return if no blocks are ever created, thus you should restrict the maximum amount of time of the test execution.

pub fn account_nonce(&self) -> u32

Get the nonce of the node account

pub fn account_nonce_of( &self, account_id: <Runtime as DomainRuntime>::AccountId, ) -> u32

Get the nonce of the given account

pub async fn send_system_remark(&self)

Sends a signed system.remark extrinsic to the pool containing the current account nonce.

pub async fn construct_and_send_extrinsic( &self, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> Result<RpcTransactionOutput, RpcTransactionError>

Construct a signed extrinsic with the current nonce of the node account and send it to this node.

pub async fn construct_and_send_extrinsic_with( &self, nonce: u32, tip: <<Runtime as Config>::OnChargeTransaction as OnChargeTransaction<Runtime>>::Balance, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> Result<RpcTransactionOutput, RpcTransactionError>

Construct a signed extrinsic with the given nonce and tip for the node account and send it to this node.

pub fn construct_extrinsic( &self, nonce: u32, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> UncheckedExtrinsic<<Runtime as DomainRuntime>::Address, <Runtime as Config>::RuntimeCall, <Runtime as DomainRuntime>::Signature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>)>

Construct a signed extrinsic.

pub fn construct_extrinsic_with_tip( &self, nonce: u32, tip: <<Runtime as Config>::OnChargeTransaction as OnChargeTransaction<Runtime>>::Balance, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> UncheckedExtrinsic<<Runtime as DomainRuntime>::Address, <Runtime as Config>::RuntimeCall, <Runtime as DomainRuntime>::Signature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>)>

Construct a signed extrinsic with the given transaction tip.

pub async fn send_extrinsic( &self, extrinsic: impl Into<OpaqueExtrinsic>, ) -> Result<RpcTransactionOutput, RpcTransactionError>

Send an extrinsic to this node.

pub fn free_balance( &self, account_id: <Runtime as DomainRuntime>::AccountId, ) -> Balance

Get the free balance of the given account

pub fn get_open_channel_for_chain(&self, chain_id: ChainId) -> Option<ChannelId>

Returns the open XDM channel for given chain

pub fn construct_unsigned_extrinsic( &self, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> UncheckedExtrinsic<<Runtime as DomainRuntime>::Address, <Runtime as Config>::RuntimeCall, <Runtime as DomainRuntime>::Signature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>)>
where Runtime: Config<Hash = H256> + Config + Send + Sync, RuntimeCallFor<Runtime>: Dispatchable<Info = DispatchInfo, PostInfo = PostDispatchInfo> + Send + Sync, <<Runtime as Config>::OnChargeTransaction as OnChargeTransaction<Runtime>>::Balance: Send + Sync + From<u64> + FixedPointOperand,

Construct an unsigned extrinsic that can be applied to the test runtime.

pub async fn construct_and_send_unsigned_extrinsic( &self, function: impl Into<<Runtime as Config>::RuntimeCall>, ) -> Result<RpcTransactionOutput, RpcTransactionError>

Construct an unsigned extrinsic and send it to this node.

pub fn ban_peer(&self, addr: MultiaddrWithPeerId)

Give the peer at addr the minimum reputation, which will ban it.

pub fn unban_peer(&self, addr: MultiaddrWithPeerId)

Give the peer at addr a high reputation, which guarantees it is un-banned it.

pub async fn stop(self) -> Result<(), Error>

Take and stop the domain node and delete its database lock file.

Stopping and restarting a node can cause weird race conditions, with errors like: “The system cannot find the path specified”. If this happens, try increasing the wait time in this method.

pub async fn clear_tx_pool(&self)

Remove all tx from the tx pool

impl<Runtime, RuntimeApi> DomainNode<Runtime, RuntimeApi>
where Runtime: Config<Hash = H256> + Config + DomainRuntime + Send + Sync, RuntimeApi: ConstructRuntimeApi<Block, FullClient<Block, RuntimeApi>> + Send + Sync + 'static, RuntimeApi::RuntimeApi: Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, <Runtime as DomainRuntime>::AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + MessengerApi<Block, NumberFor<Block>, BlockHashFor<Block>> + RelayerApi<Block, NumberFor<Block>, NumberFor<Block>, BlockHashFor<Block>> + EvmOnchainStateApi<Block>,

pub fn evm_contract_creation_allowed_by( &self, ) -> PermissionedActionAllowedBy<EthereumAccountId>

Returns the current EVM contract creation allow list. Returns None if this is not an EVM domain, or if the allow list isn’t set (allow all).