Type Alias domain_test_service::domain::EvmDomainNode

pub type EvmDomainNode = DomainNode<Runtime, RuntimeApi, EVMDomainExecutorDispatch, AccountId>;

The evm domain node

Aliased Type

struct EvmDomainNode {
    pub domain_id: DomainId,
    pub key: Keyring,
    pub task_manager: TaskManager,
    pub client: Arc<Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Global>,
    pub backend: Arc<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Global>,
    pub code_executor: Arc<NativeElseWasmExecutor<EVMDomainExecutorDispatch>, Global>,
    pub network_service: Arc<NetworkService<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, H256>, Global>,
    pub sync_service: Arc<SyncingService<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Global>,
    pub addr: MultiaddrWithPeerId,
    pub rpc_handlers: RpcHandlers,
    pub operator: Operator<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<TestExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, BasicPool<FullChainApiWrapper<Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<TestExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>,
    pub tx_pool_sink: TracingUnboundedSender<Vec<u8, Global>>,
    /* private fields */
}

Fields
domain_id: DomainId

The domain id

key: Keyring

The node’s account key

task_manager: TaskManager

TaskManager’s instance.

client: Arc<Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Global>

Client’s instance.

backend: Arc<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Global>

Client backend.

code_executor: Arc<NativeElseWasmExecutor<EVMDomainExecutorDispatch>, Global>

Code executor.

network_service: Arc<NetworkService<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, H256>, Global>

Network service.

sync_service: Arc<SyncingService<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Global>

Sync service.

addr: MultiaddrWithPeerId

The MultiaddrWithPeerId to this node. This is useful if you want to pass it as “boot node” to other nodes.

rpc_handlers: RpcHandlers

RPCHandlers to make RPC queries.

operator: Operator<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<TestExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, BasicPool<FullChainApiWrapper<Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<TestExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Client<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, LocalCallExecutor<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, NativeElseWasmExecutor<EVMDomainExecutorDispatch>>

Domain oeprator.

tx_pool_sink: TracingUnboundedSender<Vec<u8, Global>>

Sink to the node’s tx pool

Implementations

impl<Runtime, RuntimeApi, ExecutorDispatch, AccountId> DomainNode<Runtime, RuntimeApi, ExecutorDispatch, AccountId>where Runtime: Config<Hash = H256> + Config + Send + Sync, Runtime::RuntimeCall: Dispatchable<Info = DispatchInfo, PostInfo = PostDispatchInfo> + Send + Sync, <<Runtime as Config>::OnChargeTransaction as OnChargeTransaction<Runtime>>::Balance: Send + Sync + From<u64> + FixedPointOperand, u64: From<BlockNumberFor<Runtime>>, RuntimeApi: ConstructRuntimeApi<Block, FullClient<Block, RuntimeApi, ExecutorDispatch>> + Send + Sync + 'static, RuntimeApi::RuntimeApi: ApiExt<Block> + Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + MessengerApi<Block, NumberFor<Block>> + RelayerApi<Block, NumberFor<Block>> + OnchainStateApi<Block, AccountId, Balance> + EthereumRuntimeRPCApi<Block>, ExecutorDispatch: NativeExecutionDispatch + Send + Sync + 'static, AccountId: DeserializeOwned + Encode + Decode + Clone + Debug + Display + FromStr + Sync + Send + FromKeyring + 'static,

pub fn wait_for_blocks(&self, count: usize) -> impl Future<Output = ()>

Wait for count blocks to be imported in the node and then exit. This function will not return if no blocks are ever created, thus you should restrict the maximum amount of time of the test execution.

pub fn account_nonce(&self) -> u32

Get the nonce of the node account

pub async fn send_system_remark(&mut self)

Sends an system.remark extrinsic to the pool.

pub async fn construct_and_send_extrinsic( &mut self, function: impl Into<<Runtime as Config>::RuntimeCall> ) -> Result<RpcTransactionOutput, RpcTransactionError>

Construct an extrinsic with the current nonce of the node account and send it to this node.

pub fn construct_extrinsic( &mut self, nonce: u32, function: impl Into<<Runtime as Config>::RuntimeCall> ) -> UncheckedExtrinsic<Address, <Runtime as Config>::RuntimeCall, Signature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>)>

Construct an extrinsic.

pub fn construct_extrinsic_with_tip( &mut self, nonce: u32, tip: u32, function: impl Into<<Runtime as Config>::RuntimeCall> ) -> UncheckedExtrinsic<Address, <Runtime as Config>::RuntimeCall, Signature, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>)>

Construct an extrinsic with the given transaction tip.

pub async fn send_extrinsic( &self, extrinsic: impl Into<OpaqueExtrinsic> ) -> Result<RpcTransactionOutput, RpcTransactionError>

Send an extrinsic to this node.

pub fn free_balance(&self, account_id: AccountId) -> Balance

Get the free balance of the given account