Crate domain_test_service

Crate used for testing with Domain.

Re-exports

pub use keyring::Keyring as EcdsaKeyring;

pub use evm_domain_test_runtime;

pub use domain::*;

Modules

chain_spec

Chain specification for the domain test runtime.

domain

Utilities used for testing with the domain.

keyring

Set of test accounts.

Enums

Sr25519Keyring

Set of test accounts.

Constants

GENESIS_DOMAIN_ID

The domain id of the genesis domain

Functions

construct_extrinsic_generic

Construct an extrinsic that can be applied to the test runtime.

construct_extrinsic_generic_with_custom_key

Construct a generic extrinsic signed by custom key

construct_unsigned_extrinsic

Construct an unsigned extrinsic that can be applied to the test runtime.

node_config

Create a domain node Configuration.